// Codigo/Front/src/types/freightType.ts
export interface FreightConfigDTO {
  kmPerLiterConsumption: number;
  fuelPrice: number;
  maintenanceCostPerKm: number;
  tireCostPerKm: number;
  depreciationCostPerKm: number;
  insuranceCostPerKm: number;
  baseSalary: number;
  chargesPercentage: number;
  monthlyHoursWorked: number;
  administrativeCostsPercentage: number;
  marginPercentage: number;
  fixedFee: number;
}